update controllers and routes
Language refactoring in progress
