+++ src/controller/events-controller.js
@@ -4,8 +4,6 @@ const validator = require("../validator/validator");
 const authService = require("../services/auth-service");
 const md5 = require("md5");
 const connection = require("../models/init-models");
-const data = require("../validator/data");
-const arquivos = require("./arquivos-controler");
 
 const categoria = ["extensao", "iniciacao_cientifica"];
 const status_projeto = [
@@ -16,8 +14,8 @@ const status_projeto = [
 ];
 
 exports.post = async (req, res, next) => {
-  // #swagger.tags = ['Projetos']
-  // #swagger.description = 'Endpoint para cadastrar Projetos no sistema.'
+  // #swagger.tags = ['Events']
+  // #swagger.description = 'Endpoint para cadastrar Evento no sistema.'
   // #swagger.security = [{ApiKeyAuth: []}]
   /* #swagger.parameters['dados'] = {
       in: 'body',
@@ -46,25 +44,6 @@ exports.post = async (req, res, next) => {
     },
   */
 
-  if (req.body.dadosJSON) {
-    req.body = Object.assign({}, req.body, JSON.parse(req.body.dadosJSON));
-  }
-
-  if (req.files) {
-    req.body.arquivos_has_projetos = [];
-
-    const arquivosList = await arquivos.SalvarArquivos(req.files, "projeto");
-
-    arquivosList.forEach((element) => {
-      req.body.arquivos_has_projetos.push({
-        arquivo_id_arquivo_arquivo: element,
-      });
-    });
-  }
-
-  req.body.data_inicial = data.ConvertDataBRtoUS(req.body.data_inicial);
-  req.body.data_final = data.ConvertDataBRtoUS(req.body.data_final);
-
   req.body.endereco_id_endereco_endereco = req.body.endereco;
   req.body.lider_lab_id_lider_lab = req.body.jwtDecodeDados?.id_lider_lab;
 
@@ -78,55 +57,13 @@ exports.post = async (req, res, next) => {
           model: models.endereco,
           as: "endereco_id_endereco_endereco",
         },
-        {
-          model: models.arquivos_has_projetos,
-          as: "arquivos_has_projetos",
-          include: [
-            { model: models.arquivo, as: "arquivo_id_arquivo_arquivo" },
-          ],
-        },
       ],
     })
     .then(async (response) => {
       res.status(200).send({
-        message: "Projetos cadastrado com suscesso",
+        message: "Projetos cadastrado com sucesso",
         dados: {
           projeto: response,
-        },
-      });
-    })
-    .catch((err) => {
-      res.status(500).send(JSON.stringify(err?.sqlMessage));
-    })
-    .finally(() => {
-      connection.closeConnection(models.sequelize);
-    });
-};
-
-exports.getAll = async (req, res, next) => {
-  // #swagger.tags = ['Projetos']
-  // #swagger.description = 'Endpoint para listar Projetos do sistema.'
-
-  const models = connection.initModels();
-
-  models.projeto
-    .findAll({
-      include: [
-        {
-          model: models.endereco,
-          as: "endereco_id_endereco_endereco",
-        },
-        {
-          model: models.lider_lab,
-          as: "lider_lab_id_lider_lab_lider_lab",
-        },
-      ],
-    })
-    .then(async (response) => {
-      res.status(200).send({
-        message: "Projeto encontrados com suscesso",
-        dados: {
-          laboratorio: response,
         },
       });
     })
