+++ src/controller/users-controller.js
@@ -5,7 +5,7 @@ const authService = require("../services/auth-service");
 const mailService = require("../services/mail-service");
 const md5 = require("md5");
 const connection = require("../models/init-models");
-const arquivos = require("./arquivos-controler");
+const arquivos = require("./files-controller");
 const dataFunctions = require("../validator/data");
 const { Op } = require("sequelize");
 
@@ -13,11 +13,11 @@ const userTipes = ["admin", "aluno", "orientador", "lider_lab", "parceiro"];
 const addressTypes = ["fisico", "virtual", "indefinido"];
 
 exports.post = async (req, res, next) => {
-  // #swagger.tags = ['Usuario']
-  // #swagger.description = 'Endpoint para cadastrar no sistema.'
-  /* #swagger.parameters['dados'] = {
+  // #swagger.tags = ['Users']
+  // #swagger.summary = 'Insert new user into database'
+  /* #swagger.parameters['body'] = {
       in: 'body',
-      description: 'Informações para login usuário.',
+      description: 'User object that needs to be added to the system',
       required: true,
       type: 'object',
       schema: { 
@@ -144,7 +144,7 @@ exports.post = async (req, res, next) => {
     })
     .then(async (response) => {
       res.status(200).send({
-        message: "Usuario encontrado com suscesso",
+        message: "Usuario encontrado com sucesso",
         dados: {
           usuario: response,
         },
@@ -159,12 +159,12 @@ exports.post = async (req, res, next) => {
 };
 
 exports.put = async (req, res, next) => {
-  // #swagger.tags = ['Usuario']
-  // #swagger.description = 'Endpoint para atualizar dados do usuário no sistema.'
+  // #swagger.tags = ['Users']
+  // #swagger.summary = 'Update data for an existing user'
   // #swagger.security = [{ApiKeyAuth: []}]
-  /* #swagger.parameters['dados'] = {
+  /* #swagger.parameters['body'] = {
       in: 'body',
-      description: 'Dados para atualizar cadastro.',
+      description: 'User object that needs to be updated in the system',
       required: true,
       type: 'object',
       schema: { 
@@ -271,11 +271,11 @@ exports.put = async (req, res, next) => {
 };
 
 exports.login = async (req, res, next) => {
-  // #swagger.tags = ['Usuario']
-  // #swagger.description = 'Endpoint para logar no sistema.'
-  /* #swagger.parameters['dados_login'] = {
+  // #swagger.tags = ['Users']
+  // #swagger.summary = 'Generate access token with user credentials '
+  /* #swagger.parameters['body'] = {
       in: 'body',
-      description: 'Informações para login usuário.',
+      description: 'Access credentials.',
       required: true,
       type: 'object',
       schema: { 
@@ -360,7 +360,7 @@ exports.login = async (req, res, next) => {
       let jwt = await authService.generateToken(dadosUsuario);
 
       res.status(200).send({
-        message: "Usuario encontrado com suscesso",
+        message: "Usuario encontrado com sucesso",
         dados: {
           jwt: jwt,
           usuario: dadosUsuario,
@@ -377,11 +377,11 @@ exports.login = async (req, res, next) => {
 };
 
 exports.sendMailToResetPassword = async (req, res, next) => {
-  // #swagger.tags = ['Usuario']
-  // #swagger.description = 'Endpoint para envio de email com link para criação de nova senha.'
-  /* #swagger.parameters['email'] = {
+  // #swagger.tags = ['Users']
+  // #swagger.summary = 'Send an email with a link to create a new password. '
+  /* #swagger.parameters['body'] = {
       in: 'body',
-      description: 'Email para recuperação de conta.',
+      description: 'Account recovery email.',
       required: true,
       type: 'object',
       schema: { 
@@ -460,7 +460,7 @@ exports.sendMailToResetPassword = async (req, res, next) => {
       
       if (result) {
         res.status(200).send({
-          message: "Email enviado com suscesso",
+          message: "Email enviado com sucesso",
         });
       } else {
         res.status(500).send({
@@ -478,12 +478,12 @@ exports.sendMailToResetPassword = async (req, res, next) => {
 };
 
 exports.updatePassword = async (req, res, next) => {
-  // #swagger.tags = ['Usuario']
-  // #swagger.description = 'Endpoint para atualização de senha na plataforma.'
+  // #swagger.tags = ['Users']
+  // #swagger.summary = 'Update a users password with an emailed token.'
   // #swagger.security = [{ApiKeyAuth: []}]
-  /* #swagger.parameters['senha'] = {
+  /* #swagger.parameters['body'] = {
       in: 'body',
-      description: 'Nova senha escolhida para uso na plataforma.',
+      description: 'New password chosen for use on the platform.',
       required: true,
       type: 'object',
       schema: { 
@@ -527,7 +527,7 @@ exports.updatePassword = async (req, res, next) => {
       });
 
       res.status(200).send({
-        message: "Senha atualizada com suscesso",
+        message: "Senha atualizada com sucesso",
       });
     })
     .catch((err) => {
